+++ backend/canisters/community/impl/src/model/groups_being_imported.rs
@@ -1,11 +1,22 @@
 use serde::{Deserialize, Serialize};
 use std::collections::hash_map::Entry::Vacant;
 use std::collections::HashMap;
-use types::{ChannelId, ChatId, TimestampMillis, Timestamped, UserId};
+use types::{ChannelId, ChatId, EventIndex, MessageIndex, TimestampMillis, Timestamped, UserId};
 
 #[derive(Serialize, Deserialize, Default)]
 pub struct GroupsBeingImported {
     groups: HashMap<ChatId, GroupBeingImported>,
+}
+
+pub struct GroupToImport {
+    pub group_id: ChatId,
+    pub action: GroupToImportAction,
+}
+
+#[derive(Debug)]
+pub enum GroupToImportAction {
+    Events(ChannelId, Option<(Option<MessageIndex>, EventIndex)>),
+    Core(u64),
 }
 
 impl GroupsBeingImported {
@@ -31,12 +42,20 @@ impl GroupsBeingImported {
         self.groups.contains_key(group_id)
     }
 
-    pub fn next_batch(&mut self, now: TimestampMillis) -> Vec<(ChatId, u64)> {
+    pub fn next_batch(&mut self, now: TimestampMillis) -> Vec<GroupToImport> {
         let mut batch = Vec::new();
         for (chat_id, group) in self.groups.iter_mut().filter(|(_, g)| !g.is_complete()) {
             if group.current_batch_started.is_none() {
                 group.current_batch_started = Some(now);
-                batch.push((*chat_id, group.bytes.len() as u64));
+                let action = if group.events_imported {
+                    GroupToImportAction::Core(group.bytes.len() as u64)
+                } else {
+                    GroupToImportAction::Events(group.channel_id, group.events_imported_up_to)
+                };
+                batch.push(GroupToImport {
+                    group_id: *chat_id,
+                    action,
+                });
             }
         }
         batch
@@ -51,6 +70,22 @@ impl GroupsBeingImported {
             group.is_complete()
         } else {
             false
+        }
+    }
+
+    pub fn mark_events_batch_complete(&mut self, group_id: &ChatId, up_to: (Option<MessageIndex>, EventIndex)) {
+        if let Some(group) = self.groups.get_mut(group_id) {
+            group.current_batch_started = None;
+            group.error_message = None;
+            group.events_imported_up_to = Some(up_to);
+        }
+    }
+
+    pub fn mark_events_import_complete(&mut self, group_id: &ChatId) {
+        if let Some(group) = self.groups.get_mut(group_id) {
+            group.current_batch_started = None;
+            group.error_message = None;
+            group.events_imported = true;
         }
     }
 
@@ -84,6 +119,9 @@ pub struct GroupBeingImported {
     imported_by: UserId,
     import_started: TimestampMillis,
     current_batch_started: Option<TimestampMillis>,
+    #[serde(default)]
+    events_imported: bool,
+    events_imported_up_to: Option<(Option<MessageIndex>, EventIndex)>,
     total_bytes: u64,
     #[serde(with = "serde_bytes")]
     bytes: Vec<u8>,
@@ -104,6 +142,8 @@ impl GroupBeingImported {
             imported_by,
             import_started: now,
             current_batch_started: None,
+            events_imported: false,
+            events_imported_up_to: None,
             total_bytes,
             bytes: Vec::with_capacity(total_bytes as usize),
             error_message: None,