+++ backend/canisters/local_group_index/api/src/updates/c2c_mark_events_migrated_to_stable_memory.rs
@@ -1,0 +1,4 @@
+use types::Empty;
+
+pub type Args = Empty;
+pub type Response = ();