+++ backend/canisters/user/impl/src/regular_jobs.rs
@@ -1,4 +1,5 @@
-use crate::Data;
+use crate::{mutate_state, Data};
+use types::{CanisterId, Empty};
 use utils::env::Environment;
 use utils::regular_jobs::{RegularJob, RegularJobs};
 use utils::time::MINUTE_IN_MS;
@@ -11,11 +12,13 @@ pub(crate) fn build() -> RegularJobs<Data> {
         5 * MINUTE_IN_MS,
     );
     let retry_deleting_files = RegularJob::new("Retry deleting files", retry_deleting_files, MINUTE_IN_MS);
+    let migrate_chat_events_to_stable_memory = RegularJob::new("Migrate chat events", migrate_chat_events_to_stable_memory, 0);
 
     RegularJobs::new(vec![
         check_cycles_balance,
         aggregate_direct_chat_metrics,
         retry_deleting_files,
+        migrate_chat_events_to_stable_memory,
     ])
 }
 
@@ -29,4 +32,26 @@ fn aggregate_direct_chat_metrics(_: &dyn Environment, data: &mut Data) {
 
 fn retry_deleting_files(_: &dyn Environment, _: &mut Data) {
     storage_bucket_client::retry_failed();
+}
+
+fn migrate_chat_events_to_stable_memory(_: &dyn Environment, data: &mut Data) {
+    for chat in data.direct_chats.iter_mut() {
+        let (count_migrated, finished) = chat.events.migrate_next_batch_of_events_to_stable_storage();
+        if !finished {
+            return;
+        }
+        if count_migrated > 0 && ic_cdk::api::instruction_counter() > 5_000_000_000 {
+            return;
+        }
+    }
+    ic_cdk::spawn(notify_migration_to_stable_memory_complete(data.local_user_index_canister_id));
+}
+
+async fn notify_migration_to_stable_memory_complete(local_group_index: CanisterId) {
+    if local_user_index_canister_c2c_client::c2c_mark_events_migrated_to_stable_memory(local_group_index, &Empty {})
+        .await
+        .is_ok()
+    {
+        mutate_state(|state| state.data.stable_memory_event_migration_complete = true);
+    }
 }