+++ backend/integration_tests/src/p2p_swap_tests.rs
@@ -5,7 +5,7 @@ use candid::Principal;
 use std::ops::Deref;
 use std::time::Duration;
 use test_case::test_case;
-use testing::rng::{random_message_id, random_string};
+use testing::rng::{random_from_u128, random_string};
 use types::{ChatEvent, Cryptocurrency, MessageContent, MessageContentInitial, P2PSwapContentInitial, P2PSwapStatus};
 use utils::time::{DAY_IN_MS, MINUTE_IN_MS};
 
@@ -37,7 +37,7 @@ fn p2p_swap_in_direct_chat_succeeds() {
         11_000_000_000,
     );
 
-    let message_id = random_message_id();
+    let message_id = random_from_u128();
 
     let send_message_response = client::user::send_message_v2(
         env,
@@ -154,7 +154,7 @@ fn p2p_swap_in_group_succeeds() {
         11_000_000_000,
     );
 
-    let message_id = random_message_id();
+    let message_id = random_from_u128();
 
     let send_message_response = client::user::send_message_with_transfer_to_group(
         env,
@@ -254,7 +254,7 @@ fn cancel_p2p_swap_in_direct_chat_succeeds(delete_message: bool) {
         original_chat_balance,
     );
 
-    let message_id = random_message_id();
+    let message_id = random_from_u128();
 
     let send_message_response = client::user::send_message_v2(
         env,
@@ -381,7 +381,7 @@ fn cancel_p2p_swap_in_group_chat_succeeds(delete_message: bool) {
         original_chat_balance,
     );
 
-    let message_id = random_message_id();
+    let message_id = random_from_u128();
 
     let send_message_response = client::user::send_message_with_transfer_to_group(
         env,
@@ -497,7 +497,7 @@ fn deposit_refunded_if_swap_expires() {
         original_chat_balance,
     );
 
-    let message_id = random_message_id();
+    let message_id = random_from_u128();
 
     let send_message_response = client::user::send_message_v2(
         env,